Enable https proxy for api endpoints
This enables the ability to use https for the api endpoints via mod_proxy in
apache. While it appears to be a workaround, its a much easier method to enable
than doing it natively. This assumes that the native api endpoints are listening
on localhost to ensure no port conflicts.