OpenStack nagios service checks
This uses plugins from here [1] to check all various services. The current
settings are adjustable via attributes based on the cluster needs. I created an
RPM for this an uploaded it to the oslrepo so that its easier to install (it's
not on pypi).

NRPE will call sudo against a wrapper check_openstack which sources the
/root/openrc file, thus giving the check script access to the cluster. This does
pose a risk in case the plugins have a security issue however I haven't figured
out how to grant admin read-only access to an account in OpenStack. A TODO for
later would be to create a nagios specific user that only has read-only access.

These checks only get installed and setup on the controller which can check all
the other hosts.

[1] https://github.com/cirrax/openstack-nagios-plugins