Bump default_api_return_limit to see more meters
By default, ceilometer limits the amount of meters to list to 100, but when you
are checking a lot of meters, this means you don't see all of the meters. As a
workaround [1], it's suggested to bump that amount to something really high. The
unfortunate side affect is that the resources dashboard will take a while to run
depending on how much data you have, but at least you can see the meters.

[1] https://bugs.launchpad.net/ceilometer/+bug/1540204