Create openstack_ops_identity role to ease base setup
Almost every recipe requires keystone to actually test, but normally they aren't
hosted on the same machine so it doesn't make sense to keep it in that recipe.
So for now let's make a role for our testing purposes. That way if we need to
change it globally, we do it in one place and not ten.