Update the nrpe_check[check_load] resource instead of trying to override attrs
The current logic is broken in production due to attribute precedence. The
ChefSpec tests didn't catch it due to the fact that osl-nrpe wasn't included in
the converge run list. By adding it, this replicates what happens in production
and verifies this works correctly finally. I also verfied this new test fails on
the old code.