Restart linuxbridge-agent when iptable rules are updated
This works around an issue where the current setup will not apply new iptables
rules due to the that the rules are saved. Instead, use iptables like we do
elsewhere, only also restart linuxbridge-agent if rule are updated. That way
those rules are generated again. This removes the save settings in the iptables
init script so that the rules are cleared on every restart.

This uses a systemd override using PartOf which ties linuxbridge-agent to
iptable restarts. Whenever iptables is restarted, linuxbridge-agent is now also
restarted which should fill the rules that are needed properly.