Separate Network Node logic
This allows to optionally have a separate node act as a network node and run the
Neutron Server. This is going to be required for utilizing the Mellanox switch
in the future. By default the network public API will still be on the
controller, however when this new attribute is set to true, it will reside on
the new network node.

Additional fixes include:

- Update and fix chef provisioning test environment
- Add optional network node to chef provisioning test environment
- Install python-memcached so that querying neutron works globally