Retry SSH connections 3 times and also log all details in a seperate file.
SSH connections might be refused while the instance is booting up,
so keep on trying it just in case the instance is now allowing them.

Seperate log file for SSH because Net::SSH seems to accept only logger objects
and Samar is still to have better logging implemented in the whole gem.