Nagios monitoring in Newton
This pulls in nagios plugins from the OpenStack project that are being
maintained and still work in Newton. The previous plugins we were using worked
in Mitaka, but don't appear to work in Newton. This isn't exactly the same
checks as we had but it should be good enough for what we need.