Disable volumes from instance launch window in Horizon
This disables the new default setting which displays using volumes for instances
in Newton using settings found here[1]. We don't have support for this working
in our current cluster so this disables it. Unfortunately, one of the settings
(create_volume) isn't available until Pike but I'm including it in here anyway
once we get there. Uses will still need to select "No" to create volume in the
UI.

[1] https://docs.openstack.org/horizon/latest/configuration/settings.html#launch-instance-defaults