Safely shutdown and start instances on hypervisor reboots
This properly enables support which shuts down guests using the libvirt-guest
service on power down. And then enables the feature which tells OpenStack to
start guests on hypervisor boot up.

Currently this doesn't happen and we end up manually needing to start up guests
which might be in a state of corruption too since they weren't shutdown
properly.

This follows the steps described here [1].

[1] https://ask.openstack.org/en/question/66880/start-vms-after-reboot-compute-node/