Install 4.14 kernel and set kvm_cma_resv_ratio=10 on ppc64le
On ppc64le, we need to run a newer mainline kernel so that we get the best
support for the hardware. In addition, to deal with vms not booting due to CMA
(See [1] for detail on what CMA is) memory issues, let's increase the ratio from
the default of 5 to 10. This number is a percentage of memory for hash pagetable
allocation. I've been running with 10% on openpower2 for several months and it
seems to have stopped the issue.

In addition, add a new munin graph which tracks the CMA memory usage on hosts
which have that information available in /proc.

[1] https://lwn.net/Articles/486301/