Fix duplicated resources for group[ceph]
This fixes running the all-in-one (aio) suite and also resolves some other
issues where we had the same resource set in three different recipes. Even
though we're editing the same group, the resource NAME needs to be different.

In addition, we need to ensure the tempest virtualenv isn't using pip 10.0.0
which breaks it otherwise.

This should also provide some fixes for the packer builds.