Install older version of cliff python package
This resolves an issue that the current latest version has with our environment.
Unfortunately it wasn't easy to override what the upstream cookbooks were doing
so I went ahead and just replicated the code it uses and just added the one-off
case here.

Looks like this can be removed when we switch to Pike as the upstream cookbooks
move back to yum packages.