Don't use upcoming newer osl-apache (yet)
We're working on updating osl-apache to have some Chef 13 fixes and also bumping
the upstream cookbook to ~> 4.0.0 in osuosl-cookbooks/osl-apache#117.
Unfortunately this is currently incompatible with the version that is required
in the OpenStack upstream cookbooks. So to work around it, let's ensure we stay
on the 2.x branch until Pike which has the proper fixes for Chef 13.