Uninstall newer version of fog-openstack which breaks upstream openstack cb
This is a temporarly solution to get chef runs going again and points to a fork
of the openstackclient cookbook which ensures we don't install the newer
version. I'm working with upstream to get this properly fixed.

NOTE: I added the ServerSpec also in the identity suite since this doesn't get
pulled until the openstackclient gets pulled in from recipes such as identity.