Temporarily disable memcached for the metadata agent
We currently are experiencing issues where when several VMs are deployed at
once, the metadata server returns a 404 when the VMs boots up. This is a
temporary work around until we figure out a true fix. This may slow down
instance deployments a little but at least they *will* get deployed properly.