Re-enable multi-node testing / refactor upgrade suite
This re-enables and fixes the multi-node testing suite.

In addition, refactor the upgrade suite so that we do no rely on a special image
to test upgrades. Instead, allow the option of setting up the suite on the
previous release and then do the actual upgrade in the new release upgrading to.
Since this suite is rarely used this shouldn't be an issue.

Signed-off-by: Lance Albertson <lance@osuosl.org>