Refactor all code to no longer use upstream Openstack cookbooks
This pulls in all of our site logic centrally and releases all of the strange
cookbook version pinning that the old openstack cookbooks had. This also
migrates to using custom resources (where it makes sense) and cleans up the code
base quite a bit.

Another major change is moving all data from individual data bags or attributes
to a centralized data bag item per cluster.

This will make it much easier for us to upgrade the platform in the long run.

Signed-off-by: Lance Albertson <lance@osuosl.org>