Only install/use fog-openstack gem when needed
One of the gem dependencies (io-console) now requires a compiler to build a
native extension in a newer release. The compute nodes don't currently have a
compiler and it's not needed until we're using one of the resources which
requires the fog-openstack gem.

Using the logic from the postgresql cookbook, this moves the fog-openstack gem
installation out of the metadata.rb and into helper methods. If it is triggered,
then the gcc package will be installed in compile_time (not need to use
build_essential as we only need gcc), and then use chef_gem to install the gem.
This will dynamically require the fog-openstack gem if needed even after
installation.

This should fix any future issues we have with gems that fog-openstack depend on
which may also start needing native extensions.

Signed-off-by: Lance Albertson <lance@osuosl.org>