Attempt to resolve issues with snapshots
This attempts to resolve this ticket [1]. This removes the http backend with
glance so that we only have to deal with one store. This was only useful for
providing the ability of creating images from URL's. Most of our users don't
likely use this feature and it's likely causing more issues.

Additional changes:

- Remove deprecated "stores" config parameter
- Create a network and subnet and add integration tests
- Increase OSD size from 1G to 10G to allow use of Alpine image
- Add read-only access for glance ceph user to volumes and vm
- Create scripts to create network, flavor instead of creating them with chef
  directly
- Switch to using Alpine Linux image as it's known to work and is more modern
- Update tests

[1] https://support.osuosl.org/Ticket/Display.html?id=34243

Signed-off-by: Lance Albertson <lance@osuosl.org>